Update field-in-task-ref-clause.chpl for #3431
#3431 implemented an error that field-in-task-ref-clause.chpl requested.
So this test is no longer a future. I updated all the various files
to lock in the now-correct behavior.

Turns out that the test I added in #3431, errorFieldMethodInWithClause.chpl,
is in this same directory as this test.
These two are subtly different, so I am keeping them both
and adding a comment into each to point into the other.

As to the current changes to field-in-task-ref-clause.chpl:

* I fixed a typo where it said "coforall" instead of "forall".
Also enabled a 'with' clause on the forall, which was not available
at the time the test was created.

* The test used param-if folding to expose each of the 8 error lines
one at a time. Since #3431 implements the error during scopeResolve
i.e. before resolution, those param-ifs are no longer helping,
so I removed them altogether.

* Since the error in #3431 is done via USR_FATAL_CONT,
I was able to get errors for the 8 error lines all in one compiler run.

However, there was a twist... Whenever scopeResolve finds more than
1 definition of a symbol, it runs USR_STOP. (Just in case, you know.)
The way the test was written, it executed USR_STOP when finding
multiple definitions of ==, _cond_test, and =. Which prevented
more than 1 error from being reported in a single compiler run.

So I took away the conditionals and replaced assignments with writeln().
This moved me past this obstacle.
(I don't know why the compiler did not call USR_STOP for "writeln".)