Add chpl_good_alloc_size support for jemalloc
Previously we just returned minSize for jemalloc instead of asking jemalloc for
a good allocation size. Now we use `nallocx()` which "allocates no memory, but it
performs the same size computation as the mallocx() function, and returns the
real size of the allocation that would result from the equivalent mallocx()
function call".

I had been thinking that we'd have to completely switch over to the extended
API before using nallocx(), but the size classes are the same between the
standard and extended API, so it's safe to just use nallocx(). The folly
[malloc wrapper] (https://github.com/facebook/folly/blob/master/folly/Malloc.h)
does the same thing