Add a LongDouble module to LCALS
The checksums in the C version of LCALS are computed with a "long double",
which is 16 bytes with my C compiler.  Prior to this, the Chapel version had
been using a "real(64)".  For a few of the kernels the smaller precision was
not enough to get a reasonably close checksum.

Add a LongDouble module that provides a simple interface to the C type using
externs and some operator overloads.

This gets the MEDIUM and LARGE loop sizes passing verification for PIC_2D,
but SMALL is not working yet.

This gets the MEDIUM loop size passing verification for MAT_X_MAT.  The
LARGE and SMALL sizes were already passing.