Add comparison operators for int/uint pairs
While working on equality and inequality for ranges of mixed int/uint
idxType, I realized that while I continue to believe strongly that we
should not provide a (default) implementation of operators like +, -,
etc. on mixed int/uint pairs (due to the fact that in different use
cases, you might want the result to be an int or uint), this same
restriction doesn't apply to the comparison operators where the result
is always boolean.  Duh.  All that's needed is to check for a negative
int, handle that case if it is (because the answer will trivially be
true or false) and cast it to a uint otherwise, applying the normal
operator (because if it's not negative, the cast is safe).  Duh.