Fix how min/max are computed
Rather than declaring the types of min and max and assigning (which
could be unsafe in the event that we have an unbounded range as the
slicer or slicee because it will inherit the missing bound, and
therefore type, of the other range, leading to a potential type
mismatch.  See the behavior of test/types/range/vass/slice-1.chpl
prior to this fix for an example.  Instead, use a cast to make sure
the result of the min/max is the type we ultimately want.  Since
the value will be in the intersection of the two ranges, this cast
is guaranteed to be correct.