Redefine an empty, impossible to represent intersection as 1..0
In my implementation of range intersection, it seemed reasonable to
represent an empty range intersection whose natural bounds couldn't be
represented in the range's idxType as 1..0 in the slice operator.  In
doing this, I hadn't realized that this future and the spec defined
that the language should halt in this case.  Creating the degenerate
range seems much nicer than halting, and reasonably well-defined in
terms of when it will happen, so I'm proposing this instead.