Optimized comparisons on mixed int/uint pairs
With a bit more reflection, realized that I could write
these with no control flow.  It worried me for a minute
that maybe by mapping these down to a single op, we
were imposing surprising overhead (looks like it should
be one instruction, but will actually be more than one),
but given that the alternative is to insert a cast (and
given that int ops are cheap once values are in
registers compared to memory ops), I don't think this
should be a concern (who'd want a compile error +
requirement to insert a manual cast instead of these
simple ops?