Make min/max signatures more normal for base types
It seems that we've relied on a completely generic implementation
of min()/max() for some time now, which is fine when the types
match, but can result in weird conditions when the types do not
match (for example, an inability to resolve the return type).

Here, I explicitly clone min() and max() for all the normal
base types that support < and > (int, uint, real, imag) and
leave the generic versions around as a catch-all case.
I added a new test which locks in some cases that I think are
better with these new overloads.

This change was motivated by a separate attempt to fix some
min/max cases in mixed int/uint range slicing, as there was
a corner case or two which became ambiguous without these
overloads.