Re-add support for max()/min() on atomics
With the new, improved definitions of min() and max(), we can once
again support min() and max() on atomics, which removes what always
felt like a bit of a strange special case to call out specially.
When calling min()/max() on syncs/singles, it will now result in
reading those variables, just as with any other operator or function
to which they're passed when a variable of their base type is
expected.

Atomics are still special-cased to avoid getting an undefined call
to < or > for them within the definition of the generic min/max
routine.

Re-factored Vass's existing tests to combine syncs and singles
into one test that tests both min and max; moved atomics to their
own tests.