GCC 5.10 workaround
I tried to create a small C++ reproducer, but it didn't
reproduce. I was observing (in adding prints) that the for_fields
iteration would normally go over 3 fields but would stop
early if the break statement is there...

Removing the break statements lets it work again and only has
a minor compile performance cost. It should have the same
functionality.