Update .good file for readstring to use bytes in text mode too
It already used a byte count in binary mode.
And readstring is documented as taking in a number of bytes.

In the future, we might want to have an additional argument
to specify the number of Unicode code points to read, but
that's subject to our general Unicode design choices.