Adjust channel.readbits/writebits to accept other integral types
Originally, these functions insisted on uint for the value
to read/write and int(8) for the count of bits. That required
some awkward casting in the Mandelbrot MPlot.chpl for tutorials.
I've had a TO-DO for some time to remove the need for the casts.

First, this change adjusts the value and numbits arguments for
channel.read/writebits to accept any integral type.

Then, this change also adds new checks. It uses the same config param as
safeCast to skip extra checks with --fast. If the checks are not skipped, they
generate a halt if there is an attempt to read a negative number of bits or to
read more bits than the passed value. (E.g. reading 9 bits into a uint(8)).

Lastly, this change adds tests.