Allow named reduce intents i.e. forall reduce intents for min, max and user-defined reductions.
COMPILER CHANGES

Modified the parser to allow:

  ident 'reduce' ident

as an intent in a with-clause. To avoid shift-reduce conflicts,
disabled default intents (just ident), as suggested by Brad.

Convert "max" to "MaxReduceScanOp" for reduce intent representation,
just like it's done for reductions in buildReduceScanPreface1().

I verified that min/max reductions already undergo conversion to foralls.

I will commit the modified bison-chapel.cpp separately.

WHILE THERE

Rephrased the error message "can't use ... intent in a 'with' clause"
so it is readable. Also made the compiler survive (not crash) when it occurs.

Changed two possibly user-facing strings from "blank intent" to
"default intent" because that's what it is now called.

TEST CHANGES

Added a test of reduce intents for min, max and a user-defined reduction.
Added a test for the illegal intent error message.