Fix a nondeterminant test.
This test depended on sleep() to cause it to produce dependable output.
However, that doesn't always work, because sleep isn't perfectly
dependable.  Here, rewrite the test to say whether or not we're serial
(and thus expect dependably ordered output) or not in the output.  Then,
process the output before diff'ing such that the output produced when
parallel is just checked for the right number of each message, while the
output produced while serial is compared exactly.

This fixes https://chapel.atlassian.net/browse/CHAPEL-169.