Create LocaleModelHelp modules
In preparation for creating more locale models, I'm moving most of the code
that really has to do with interacting with the Chapel runtime (vs modelling a
specific architecture) into locale model help modules.

I might add 2 more locale models, and it would be a shame if we had
to update 4 files if the runtime interface changed.