Refactor standard layout/distribution documentation
This mod removes the separate documents for layouts and
distributions wcich seemed to add little value beyond
the top-level layoutdist.rst file.  BenA indicated on
email that he wanted to do something like this as well,
but couldn't make the sidebar look the way he wanted.
To make this work, I changed the top-level heading from
a '------' style to a '=====' style, which is what we
use in the user's guide as well to get a similar effect,
and I don't see any particular downside.