Clean up unnecessary pieces of the vector class in LCALS
Most of the uses of this class have been replaced with enumerated arrays
or arrays with sizes known at their declaration.  There is only one variable
left using this class that can't be easily replaced. It would be nice to use
an array-as-vec style array directly instead of the vector wrapper class for
this variable, but arrays of arrays with each inner array having its own
domain (e.g. skyline arrays) is not currently supported.

Now that there's only one variable minimally using the class, several of the
methods could be removed.