Fix gcc 5.1.0 gen compiler ABI issues
For our binary release on Cray systems we build Chapel with a "gen" compiler.
We recently upgraded our gen compiler from 4.8.2 to 5.1.0, but we encountered
several issues when intel is the target compiler. While building the runtime
and third-party code for the intel configurations we have gcc/5.1.0 and a
recent version of intel loaded. However, users can have different (and possibly
older) versions loaded when the invoke chapel themselves.

The main problem stems from the fact that gcc 5.1.0 introduced a new ABI in
order to be C11 compliant, but you can't link code that's not using the new ABI
with libs that were built with it. i.e. If we build re2 (our only c++ code in
the "runtime") with gcc 5.1.0 and a users tries to use an older version,
they'll have an ABI mismatch and get errors like "undefined references to
std::__cxx11...".

To resolve this, we now define _GLIBCXX_USE_CXX11_ABI to 0 in order to force
our module build to use the old ABI. This way users can either use the old or
the new ABI and they'll be able to link with what we generated. See
https://gcc.gnu.org/onlinedocs/libstdc++/manual/using_dual_abi.html for more
info.

gcc 5.1.0 also defaults to `-std=gnu11`. gcc 4.9 introduced the out_of_range
exception. If we build re2 with c++11 and then try to link to it with an older
gcc we get undefined references to std::__throw_out_of_range_fmt. re2 doesn't
actually use exceptions, so this just disables exceptions so that we don't have
references to the out_of_range exception.