Fix a syntax error in this future (won't resolve it)
A user's question about first-class functions made me take a look at this
related future - it turns out the future has been failing in a different way
due to a replacement error I believe when the numPUs change went in.  The
behavior of this future still doesn't fully exercise what it intends to, but
I don't have a firm enough grasp on the Cyclic distribution to off-hand know
what to do next.  But at least the syntax error is gone!