Support lldb more explicitly; don't assume xterm path.
(Theses changes reflect very helpful review feedback -- thanks!)

The previous work assumed that Mac users would always want to use lldb.
This isn't the case; some will want to use gdb instead.  Here, add
support for a CHPL_COMM_USE_LLDB environment variable to specify the use
of lldb instead of gdb.  A programmer selects which debugger to use by
setting either the old CHPL_COMM_USE_GDB environment variable or this
new one.

Also, replace the use of fixed paths for the xterm executable with a
PATH-based search analogous to the one a shell would use.  (We don't
have to do the same for the gdb or lldb executable because the shell
running in the xterm will handle that.  But GASNet will use an execve(2)
to run the xterm, and that only accepts a full path; it doesn't do a
search.