Initial implementation to make chpldoc generate one file per module
Probably needs to be cleaned up.  Also, there will be chpldoc tests that should
adjust, I haven't updated their output yet.

One issue not yet addressed - currently, the strategy to link the submodules in
to the parent module ends up also bringing them into the sidebar at the top
level.  This is undesireable, we would like to only have them visible under
the parent module and in the module index.  Trying to fix this issue is why
I have saved the code as it is now, so that I have a rollback point.