Additional clean up to a first class functions future
It turns out the reason we were getting a different error message than the one
expected by the future was because we were using an old syntax for declaring
a domain.  With this change, the message generated now matches the old expected
internal error.  Add that internal error message as a .bad file for the future,
so it doesn't slip again.