Remove NPB IS as a performance test
Performance of this test has always been... lacking, to say the least
and a core member of the team has never had the opportunity to dive
into it and see what's going on.  Meanwhile, ISx (test/studies/isx and
test/release/exampes/benchmarks/isx) is taking over our interest from
an IS perspective as a more modern take on the classic, so we're more
likely to put future effort into its performance.  To that end, remove
IS as a performance test, but keep it around as a correctness test.