Add parallel and vectorizeOnly variants of PIC_2D kernel
Make copies of the serial PIC_2D kernel, changing "for i in 0..#len" into
"forall i in 0..#len" or "forall i in vectorizeOnly(0..#len)".  The OpenMP
reference version of PIC_2D uses "#pragma omp atomic" to update an array 'h'
in the parallel loop atomically.  In these new variants, used an array of
atomics that is initialized with 'h' before the main loop, then copied back
to 'h' after the main loop.