Split modules/standard into modules/packages
Based on recent discussion, we've split modules/standard into
modules/standard and modules/packages to differentiate between
those modules that we consider should be part of any standard
Chapel implementation (modules/standard) vs. those that we
happen to bundle either for convenience or because they're
not yet mature enough to be considered part of (even a draft of)
the standard modules.

Update README to reflect this change as well as others and to
clarify the roles of the different module paths.

Update the compiler to automatically search modules/standard
automatically.