Fix deprecation of AdvancedIters
This adds an "AdvancedIters" module back in to help people know
that the name of the module has changed and adds a test to
remind us to remove this module after the release.

It also fixes the name of the module defined by DynamicIters
to match the filename and renames a debug config param to
match the new module/file name.