Here's my proposed fix to the Random sidebar issue
Lydia: I think the problem was the ** (which seems to mean recursive
glob search?) rather than anything deeper.  Simply removing the **
creates a problem because SysCTypes is stored at a deeper level and
we'd like it to sort naturally with the others.  But I couldn't see
a good reason that it was stored at that deeper level in the doc
tree (particularly if we're redefining level as "module nesting" in
this context), so why not just hoist it to the top?