Brad's fix to the Random sidebar issue
The problem there seemed to be the ** (which seems to mean recursive
glob search?) rather than anything deeper.  Simply removing the **
created a problem because SysCTypes is stored at a deeper level and
we'd like it to sort naturally with the others.  But it didn't need
to be at that deeper level in the doc tree (particularly if we're
redefining level as "module nesting" in this context), so we just
hoist it to the top.