Testing updates
Now that submodules are given their own files, we need to cat those files as
well (and adjust for the different indentation and naming of the submodule).

This also resolved the future
chpldoc/compflags/folder/save-sphinx/deepNestModules.doc.chpl, where Sphinx
would misassociate top-level functions that were listed after a submodule to the
submodule rather than the parent module.  Remove the .future and .bad and
update the .good for the new expected output.  Also update the prediff to find
the submodule file as well.