Adding missing directories to clean to Makefile
Lydia and I just noticed that when I added modules/packages/ to the
things to be documented yesterday, I failed to add that directory
to the list of things to be cleaned up.  Turns out that previous
additions of dists and layouts were never added either.  Add all
these and re-alphabetize.