Fix bug introduced by moving the doc version of SysCTypes
Moving that generated file to be a sibling to the other standard modules caused
build ambiguity, because now we have two options for SysCTypes (though only one
good choice).  Clean up the generated .chpl file for SysCTypes once we're done
with it.  Note that this means that we will run into an error if we ever decide
to drop the actual module SysCTypes.chpl as a sibling to the other standard
modules.