Adjust CHPL_RT_NUM_THREADS_PER_LOCALE messaging.
If we can't parse the CHPL_RT_NUM_THREADS_PER_LOCALE value, make the
message an error, not a warning.  Add a test to lock in this behavior.

For CHPL_TASKS=qthreads, if we reduce the CHPL_RT_NUM_THREADS_PER_LOCALE
value so as not to oversubscribe the PU resources, say so by default,
not just when verbose.  Also, don't print messages about one environment
variable overriding another if the values are the same.

Adjust tests that set very high CHPL_RT_NUM_THREADS_PER_LOCALE values to
ignore the resulting message about reducing that.