Fix gcc 5.1 ABI issues
For our binary release on Cray systems we build Chapel with a "gen" compiler.
We recently upgraded our gen compiler from 4.8.2 to 5.1.0.

gcc 5.1.0 introduced a new C++ ABI in order to be C11 compliant, but you can't
link code that's not using the new ABI with libs that were built with it. i.e.
If we build re2 (currently our only c++ code in the "runtime") with gcc 5.1.0
and a users tries to use an older version of gcc, they'll have an ABI mismatch
and get errors like "undefined references to std::__cxx11...". This can be
reproduced by compiling the runtime/library with 5.1, swapping to 4.9 and
trying to compile release/examples/patterns/recordio.chpl (or any test that
uses re2.)

To resolve this, we now define _GLIBCXX_USE_CXX11_ABI to 0 in order to force
the use of the old ABI. This way users can either use the old or the new ABI
and they'll be able to link with what we generated. See
https://gcc.gnu.org/onlinedocs/libstdc++/manual/using_dual_abi.html for more
info on the ABI change.

Note that there is probably a more elegant/principled way of doing this, but
for now I'm just adding a CXXFLAGS variable with -D_GLIBCXX_USE_CXX11_ABI=0 and
adding CXXFLAGS to the re2 makefiles. Since it's currently our only
runtime/third-party code that uses C++ this works, but if we add more in the
future we probably want to be more principled with how we set CXXFLAGS