Document our ability to use enums in the spec
Basically, moved the description of the use of modules from "Using Modules" to
"The Use Statement", in the process alterring much of the phrasing to refer to
modules or enumerated types instead of just modules.  I left one or two bits
which were only relevant to modules in "Using Modules" and gave "The Use
Statement" a reference to that section so the information wouldn't get lost due
to the shuffle.  Also, reference the ability to use enums from the enum
description section, "Enumerated Types".