Add two new chpldoc tests, of param and type methods
The param method, or a method that can only be called on a param instance of
that type, gets output correctly by chpldoc.  The type method, or a method that
can only be called on the type itself rather than a particular instance of that
type, currently does not work.

Neither of these generate correct html output.