Fix bug with generated textual chpldoc output for type methods
Turns out that AstToText was keying off the intent of the this argument in the
function, but in the case of a type function, the intent on the this argument
was blank.  Happily, the FnSymbol has a field "thisTag" which corresponds to
whether the method is a param or type method, so I updated the appendThisIntent
method on AstToText to use that instead.  This seems much cleaner, and the
tests work now.

Still need to fix the sphinx output for these cases.