Allow our venv to use the system python
Add a helper script and makefile target to force our virtualenv scripts to use
the system python. For the binary release of the cray module we need to be able
to support multiple pythons. In particular, the same module will be used on a
system with 2.6 and another with 2.7.

However, scripts installed in the virtualenv bin/ dir specify which python
version they were built with and the virtualenv includes a python that kind-of
symlinks to the system python used to initialize it. This means that the python
the virtualenv was built with needs to be available to use it. We build the
virtualenv with 2.6, but some machines only have 2.7, so we need some way to
remove the dependency on the system python version.

This adds a somewhat lame version of making the scripts in bin/ python
agnostic. It replaces `#!/usr/bin/env python<version>` with `#!/usr/bin/env
python` and removes the python* bins in the virtualenv.

I think all we'll need to do for the module build is build the venv with 2.7,
then force re-build it for 2.6 and then run this script and things *should*
work. It's not a particularly elegant solution. I don't know anything about
distributing python code in binaries so somebody else can come up with a better
solution later. In the meantime this should allow the module start_test and
chpldoc to work on CLE 5.X and 6.X