Update LICENSE file WRT what packages are used by default
The LICENSE tries to express when certain third-party packages are used.
Previously it tried to give a sense of when we default to using certain
packages (i.e. it said that we used qthreads most places except on cygwin/knc.)
This information is out of date for this release. Instead of updating it and
trying to maintain yet another list of when certain third-party packages are
used I updated the file to basically say check your printchplenv output to
determine if a package is being used.