Fix chpldoc cross-references in Memory and Time.
This mostly removes :type: roles for fundamental types, but in a few
other cases it adds :type: roles for types where we have a definition we
can point the cross-reference to.  And in a couple of "while I was here"
cases, it adds back-tics around fundamental type names that had neither
those nor :type: roles but instead we just in plaintext.