Correct line number with domain queries.
We used to generate an incorrect line number in this case:

    proc foo(X: [?D]?T) {
      ... D ...
    }

The SymExprs for the uses of D within foo() referred to
the line number with "proc foo" instead of the line number
where those uses actually occurred.

Fixed by adding SET_LINENO() appropriately.

I verified that I also fixed the line number for the uses of T
within foo(). I have not added a test for this because
I do not know how - contributions are welcome here.

I am de-futurizing this test, which produces the expected outcome now:

  test/domains/bradc/modDomQuery.chpl

It was really fixed in #3548. The error message had an incorrect
line number there, which is fixed in this change.

Thank you Lydia for bringing this to my attention.