Update license file to point to chplenv
Lydia pointed out that my previous change meant people would have to build
Chapel to get a sense of which licenses they would be using, but that's not
helpful to non-programmers.

Now we link people to chplenv so they can figure out which packages are used by
default and which are built speculatively so they can get a sense of which
licenses are used for a typical installation.