Don't under-sleep.
The qthreads and fifo versions of chpl_task_sleep() could sleep the
calling task for less than the desired amount of time.  Qthreads could
only do this in a task not bound to a shepherd, meaning early during
runtime initialization, so that was a fairly minor failing.  But fifo
could do it any time.  Adjust both of these so that they sleep the
calling task at least as long as requested, yielding while doing so.