Move the submodule output to the top of the file.
Brad pointed out it was nicer to have the submodules listed at the top of the
file if they exist rather than have them listed at the bottom of the file.  This
means an extra pass through the module's contents to look for submodules, but
since we only need to know of one, we don't have to pay this entire cost all
the time.  Move the table of contents generation to occur between the module
name output and that of its documentation comment, to avoid the submodules
getting lost in the middle.