Added chpldocs to individual functions in UtilMisc_forDocs.chpl
The main point of this change is similar to #3607:
We used to have a single chpldoc comment for a group
of functions - in this case compilerError, compilerWarning,
compilerAssert. To the reader of our online docs this looks
like incomplete documentation. So for now, short of a better
solution, like #3607, I am adding a comment for each of these
functions individually.

I wanted to aggregate compilerAssert into just a couple
of nice-looking prototypes and call the int argument 'errorDepth'
- in this file - so that the online docs look pretty.

I chose against it - to avoid situations where the user invokes
compilerAssert() according to one of these prototypes and gets
an "unresolved call" error.

So we really should do the aggregation in ChapelBase.chpl first.
See TODO in UtilMisc_forDocs.chpl for further discussion.

While there:

* Put the version without errorDepth before the version with errorDepth
for each of compilerWarning and compilerError, and edit their chpldoc
comments, to my taste.

This does not count as "code change" because this file is not used
when compiling any Chapel programs.

* Leave a single min() and max() prototype each.
Also, do not reflect the several min/max implementations on numeric types
that have been added in ChapelBase.chpl.

On the one hand, this simplification may mislead the user in some
corner cases. OTOH it gives the user easier-to-read documentation.
Discussing this with Brad, we agreed that the former is not very
likely and is worth the benefit the latter.

Other TODOs - first in ChapelBase.chpl, then reflect in UtilMisc_forDocs.chpl:

* Switch from c_string to string argument types in compilerError
and compilerWarning.

* In compilerAssert, either require the formals to be 'string' or cast
them to 'string' when passing to compilerError.