Add test filenames to users guide code samples
This commit uses the :caption: option and some symbolic link
magic to make the code block examples in the user's guide
point to the file from which it originated.

One hitch is that if I use the same .chpl file twice and
use the :caption: option for both, I get an rst-level
error.  One workaround is to supply a non-default caption
for the second/subsequent cases.  Instead, here, I've
decided to not label continuations and hope that that's
clear from context.  I'm open to input / other ideas here,
though.

The formatting of these captions is somewhat funky on my
home Mac -- there's a straggling paragraph marker which
seems to be doing what the little chain link icon normally
does on our docs pages.  I'm hoping this is a personal
problem, but not optimistic.