future: casting large uint(32) through string on ILP32
When the underlying C sizeof(long) is 4, a string representing an
integer between LONG_MAX and ULONG_MAX will be clipped to LONG_MAX
when cast to uint(32).  This is due to strtol() in the casting code.

I could only test on cygwin32 (so that's what the skipif restricts it to),
but I expect this will fail on linux32 as well.