Fix an "undefined font" warning when building spec
Recently I started getting this warning when building the spec on my
Linux desktop:

  LaTeX Font Warning: Font shape `OMS/ptm/bx/n' undefined

It is due to the following piece:

  $\backslash$\$\hspace{5pt}$

that got added to Lexical_Structure.tex in 125b7fb in #3081.

I replaced that piece with:

  $\backslash$$\backslash$$\hspace{5pt}$

and the warning went away.

I could probably get rid of the internal $$.
I put them in to keep the structure of that line more uniform.