Remove program error bit in ref return intent section
(Based on Mike's feedbock)

It may or may not be reasonable to bring this sort of
thing up directly in the specification, but either
way if we do bring it up, ref return intent is only
one of several ways to get into trouble. Thus, it
should be discussed somewhere else in the spec if
at all.