One option for the chpldoc "parent output" attempt
Partial implementation.  In html/rst mode, will create a "Usage" title and
output a code block with "use <moduleName>;" as its sole contents.  No links are
generated.

TODOs for this branch:
- Make "use" bold in that block by having the code block interpreted as chapel
  code
- Make output for txt mode

I am making an alternate branch to try and generate links for all the parent
modules (and the current module for a uniform look).