Get chpldoc to output the usage information for a given module
In html/rst mode, will create a "Usage" title, and a separate line with a
description of the use statement required to access this module, with links to
all modules in the chain.  In txt mode, prints "Usage:" with a buffer from the
module name, following by the unaltered use statement.

Still need to update the tests.