Updating bugs.rst for the release
I forgot that this is embedded in the release's documentation.  Given
the number of developers who are willing to get commit privileges, I
thought it would be useful to add a section pointing to the test
system and how to file a future.  While here, I did a little
wordsmithing as well.