Fix one last under-sleep possibility and re-enable sleepTimeUnits.
The tasking layers would under-sleep by 1 microsecond if, when computing
the (integer) microsecond part of the struct timeval for the deadline to
awaken, they truncated instead of rounding.  Here, fix this by rounding
explicitly.

In addition, re-enable the sleepTimeUnits test that we disabled due to
intermittent failures when it was new.  However, make it complain only
when we under-sleep (with zero tolerance).  Checking for over-sleep
isn't worthwhile, because that can happen for reasons that are beyond
our control.