Move the note about Assert and Math being used by default closer to the top
Ben pointed out that if we can't easily remove the "Usage" section for modules
that are included by default (Assert, Math, Types) we should at least make their
note about default usage live closer to the top.  We don't want the note to be
grabbed as the summary for the module, so put it after the first line synopsis.
The Type module's comment didn't need to be updated, since the synopsis is the
only other line in the top level module comment.