At BenA's request, made Makefile targets in doc/sphinx self-hosted
Prior to this commit, top-level sphinx-related Makefile targets
invoked the run-in-venv script when building targets in the doc/sphinx
directory, but making targets in doc/sphinx did not.  For convenience,
this commit uses run-in-venv within doc/sphinx/Makefile as well.  It
also adds a 'checkdocs' target to that Makefile to better mirror the
top-level target.