Filter out apsys errors
We've been getting some sporadic failures along the lines of:

    aprun: Unexpected close of the apsys control connection
    aprun: Exiting due to errors. Application aborted

with some of our XE testing. This adds filtering to sub_test to pick out these
errors and report them with reference to the JIRA issue so they're easier to
pick out in the regression mails without having to go dig through the nightly
logs.