Allow umask test to generate reliable output
This test verifies that updates to the umask in one task affect other tasks on
that same locale.  On cygwin, we witnessed a case where multiple tasks saw the
same original value, leading to a failure with this test.  That is because umask
is not guaranteed to be thread-safe.  To get around that, force the independent
tasks to operate serially.  This accomplishes the original purpose of the test
without leading to non-deterministic output.