Fix a clang warning in generated code
David pointed out that commit 88e5b21d00124aa236c908c224348b532e232d74
introduced a warning in Chapel-generated code for programs as simple
as test/release/examples/hello4-datapar-dist.chpl.  Specifically,
'chpl' generated a version of dsiReallocate() that would call itself
unconditionally for the 'else' branch of dsiReallocate() in
DefaultRectangular.chpl.  This seems to have been due to the way I
was testing to see whether the 'd' argument was defaultrectangular
or not, by comparing its type to 'dom's type.

In this commit, I take a different approach that I thought about but
didn't pursue originally, which is to query more directly whether the
domain was a default rectangular.  We haven't supported such a query
in the past (and I don't think it can be done via a simple type
comparison since DefaultRectangular is generic), so here, I add the
same type of query that we already support on the
DefaultRectangularArr class by adding a param method to BaseDom and
DefaultRectangularDom indicating whether or not it's a default
rectangular.  While this is somewhat of a special case (i.e., we
wouldn't want to do it for all layouts and distributions), default
rectangulars are pretty special semantically, so it seems justifiable
(and I've found myself wanting it in other cases where I was
disappointed to find that only the arrays support such a query).